requirements.txt
Flask==3.1.2
flask-cors==5.0.0
faiss-cpu==1.13.1
sentence-transformers==5.2.0
torch==2.9.1
torchvision==0.24.1
open_clip_torch==3.2.0
transformers==4.57.3
Pillow==11.0.0
numpy==1.26.4
gunicorn==23.0.0
